requirements.txt
python-telegram-bot==21.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
